added the img in the investin restaurant part
